Add org-aware email template resolution to support school-specific Resend templates. Introduces resolveTemplateId() with a registry-driven override map, adds UT variants for all 5 email types, and refactors the mutual match and message notification emails from inline API route code into the typed catalog system.
